requirements.txt
# Web框架
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# 数据库
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
SQLAlchemy==2.0.23
alembic==1.12.1

# 用户认证
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# 前端
Flask-Bootstrap4==4.0.2
Flask-Moment==1.0.5

# 工具包
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
cryptography==41.0.7

# 区块链相关
web3==6.11.1
eth-account==0.9.0
eth-utils==2.2.0

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# 生产环境
gunicorn==21.2.0
gevent==23.9.1